amd/stoneyridge: Fix pm_read16 argument size
Make pm_read16() consistent with the other PM register access functions.

Reviewed-on: https://review.coreboot.org/21748
Tested-by: build bot (Jenkins) <no-reply@coreboot.org>

+++ src/soc/amd/stoneyridge/southbridge.c
@@ -53,7 +53,7 @@ void pm_write16(u8 reg, u16 value)
 	write16((void *)(PM_MMIO_BASE + reg), value);
 }
 
-u16 pm_read16(u16 reg)
+u16 pm_read16(u8 reg)
 {
 	return read16((void *)(PM_MMIO_BASE + reg));
 }
